requirements.txt
fastapi[all]
sqlalchemy
alembic
psycopg2-binary
pydantic[email]
pytest
httpx
python-dotenv
